All lint errors fixed! 🎉 ...
- moved mapStateSelector getters to new file
